(1) In the CmlJobManager class, the method CalcTaskInputParam was modified so that it does not randomly permute the SNR points in different tasks. Instead, it puts the unfinished SNR points sequentially at the beginning of the SNR vector and divides the number of required frame errors AND trials to the number of generated tasks.
(2) The methods required for PLOT operation in the CmlJobManager class were developed and its DetermineStopFlag method was modified to save pdf figures associated with each job in the Figures directory.

(3) The JobManager class was modified so that it generates less messages when an active user does not have any JobIn, JobRunning, or TaskOut files.